Fix Parquet plugin build on macOS

+++ plugins/parquet/parquet.cpp
@@ -465,7 +465,7 @@ public:
       co_yield slice;
   }
 
-  [[nodiscard]] size_t num_events() const override {
+  [[nodiscard]] uint64_t num_events() const override {
     return num_rows_;
   }
